requirements.txt
streamlit==1.16.0
opencv-python==4.5.3
Pillow==8.2.0
numpy==1.21.0
